Remove ClassItem.hasPrivateConstructor
It is not needed as it is used in the following expression.

    cls.hasPrivateConstructor || cls.constructors().isNotEmpty()

In order for hasPrivateConstructor to be true the list of constructors
must be non empty so the second expression will always be true if
hasPrivateConstructor is true.

Test: ./gradlew test && TH
Change-Id: I2d339b8f2a50f440cd77f5b18a2d5fae505e59a4
(cherry picked from commit 4eb0db40ac3466dca1c2a38a0cbce9b11712d7ed)