Preparing to merge metalava build tasks
Currently, during the build, for each API surface (public / system / test / etc), we run metalava
four times (i.e. with --api, --api-lint, --check-compatibility:*:current and
--check-compatibility:api:released), even though they all take the exact same set of source files.

Because parsing the source files is the slowest part of metalava, we want to merge them into
one metalava invocation, which metalava actually supports already -- one can just pass all
these arguments at once, and metalava run them sequentially. Each task is actually pretty fast,
so we don't even have to run them in parallel.

However, the problem is that we'll have to different values to some of the arguments for different
tasks -- for example, we'll need to pass different --baseline and --update-baseline to api-lint,
and the released-API check.

Also, currently when metalava finishes, soong detects it from the process status code, and
show different messages for different tasks, which will no longer work once we merge the tasks.

So, this CL introduces several set of (even more) new flags:

- Task specific baseline flags
    --baseline:api-lint FILE
    --update-baseline:api-lint FILE
    --baseline:compatibility:released FILE
    --update-baseline:compatibility:released FILE

These are the same as --baseline and --update-baseline, but used only for api-lint
and the released-API check.
Note the current-API check should never be used with a baseline file, so we don't have
a baseline for this.

- Task specific error message
    --error-message:api-lint MESSAGE
    --error-message:compatibility:released MESSAGE
    --error-message:compatibility:current MESSAGE

When specified, metalava shows these messages when the corresponding tasks fail.

- Internally, we now generate multiple "Reporter" instances and give them different baselines
and different messages, and pass them to specific operations.

Bug: 151160048
Test: ./gradlew
Test: build (tree hugger)
Change-Id: Icd93fcee05fe7bde317de373ad7260e945216b9a