[Metalava] Add "strict input" mode
By default, metalava infers the source file location and tries to find missing
source files that are not specified in the command line, which is a problem
with incremental builds.

Added a few options to control this behavior:
--strict-input-files <file>
Do not let metalava read files that are not explicitly specified in the
command line. All violations are written to a given file.

Note this only applies on *read*, but not write. All paths specified in the command line,
both for reads and writes, are allowed.

When a directory is specified in the command line, all the files under this directory
will be whitelisted.

Reads on directories that are not explicitly specified are logged, but allowed.

--strict-input-files:stack <file>
Same as --strict-input-files but also print stacktraces.

--strict-input-files-exempt <file or directory>
Used with --strict-input-files. Explicitly allow  access to a given path.

When a violation is detected, metalava simply logs it in the file and
continues operating, but it finishes with a failure status code.

Also add --no-implicit-root to disable implicit root detection.

Bug: 154866080
Test: Manual test with the "api-stubs-doc" metalava command with
"--strict-input-files:warn /tmp/violations.txt" and -sourcepath ''

Change-Id: I3e9c565460f1b370cb89d73cf76204ddba27fa4a