Treat objects and final fields as non null
Companion objects in Kotlin, as well as other objects, should be
treated as non null.

Similarly, if you have a final field in Java that is initialized
to something which is clearly non null (such as the result of a
method call or a field reference that is annotated with @NonNull)
then treat it as non-null as well.

Test: Unit tests included
Bug: None filed
Change-Id: I9fdce29f2c77ba811cd23b68c2c72ebf312a960c