Don't flag missing nullness information in Kotlin enums
The PSI elements for the synthetic methods generated for enums in
Kotlin are missing nullness information. This CL works around this
such that it doesn't show up as platform types in signature files,
such that it doesn't fire MissingNullability warnings, etc.

This CL also cleans up the handling a bit for these methods by
introducing (a) a explicit synthetic property on items, and (b) a
dedicated lookup method to see if a method is one of the synthetic
enum methods, and updates existing code to use these new methods
instead of the previous tedious way of checking names, parameters
and containing class. This also revealed an existing bug where
overloading the values() method in enums did not add the overloaded
method into the stubs.

Test: Updated unit tests
Change-Id: Ia7a4dce0a77f5a47b658fb888bcc41f0b92f5b24