Moves to V4 signature files and changes default value tracking behavior
In V4+, defaults are tracked in a concise manner: instead of recording the body of the default value, we now only track the presence of a default value with the `optional` keyword placed before the rest of the parameter's modifiers.

Making a default parameter non-default is a breaking change to source & binary compatability, so it is important to track this - but the actual body of the default does not matter for compatibility.

Bug: b/145971821
Bug: b/168728820
Test: ApiFileTest

Change-Id: I9292ba50f3eecbff389799af189cfffb8c5bf62b