Fix metalava treatment of "for stubs" annotations
Previously, if an API was annotated with multiple showAnnotations
but also one showForStubsPurposesAnnotations, it would be excluded.

The correct behavior is to include the API if it is annotated with any
showAnnotation that isn't also a showForStubsPurposesAnnotation.

Bug: 170447778
Test: ShowForStubPurposesAnnotationsTest
Test: exp. CL with -showAnnotation TestApi -showForStubs SystemApi
Change-Id: Iac1d6ae6bd544db25c26f64f50e4c4df787dc9f6
Merged-In: Iac1d6ae6bd544db25c26f64f50e4c4df787dc9f6