Fix conflicting API lint checks builder units
This change enables builders to have chainable setters for fraction or
percentage properties checked by `FractionFloat` or `PercentageInt` such
as `public MyBuilder setMyFraction(float f)`.

It accomplishes this by reducing the scope of these checks to validate
that the method type is not the banned numeric type. That is,
`FractionFloat` validates that the type is not an int, long, or short
and `PercentageInt` is not a float or double. This brings it more in
line with the behavior of the `MethodNameUnits` check.

Test: ApiLintTest
Fix: 174188159
Change-Id: Ib423ea40c93f85360227abd9825e95eed881354e