Relax AllUpper check on non-const Kotlin vals
The Kotlin coding conventions suggest using lowerCamelCase for mutable
properties or constants containing mutable objects, ALL_UPPER_CASE for
constants, and UpperCamelCase for references to singletons or objects.
<https://kotlinlang.org/docs/reference/coding-conventions.html#property-names>

It is infeasible to determine if a constant field is a reference to a
singleton or a mutable object from the ApiLint context. But it is known
that all `const val` properties should be in ALL_UPPER_CASE. This change
only enforces the AllUpper check on `const val` properties, leaving the
case of other properties to the discretion of the developer.

Test: ApiLintTest
Fix: 135708486
Change-Id: Icb65ddf2ccef6eedd0b9dbbdbb723f5d21b41c7a