CompatibilityCheck implicitly adding public apis to its filter
even when --show-annotation is present. Even if a caller passes --show-annotation to refer to the set of APIs annotated with a specific annotation, they really mean the set of APIs having that annotation plus the set of public APIs.

This facilitates a subsequent change which instructs CompatibilityCheck to merge multiple codebases before comparing them, where moving an api from public (unannotated) to SystemApi should still be considered part of the same codebase.

This change is extracted into its own commit so that it can be validated separately and not necessarily need be reverted if a subsequent change ends up causing unexpected failures.

This change doesn't add any addition tests because it's not expected to really be a behavior change on its own.

Bug: 181972033
Test: ./gradlew test # runs existing tests

Change-Id: I1ad0ea81163036b76e7eed172e57874906ca3b91