Include SdkConstant in jar stubs
Writing annotations with source retention into the jar stub source code
is normally not meaningful because they will not be visible to users of
the stubs. For modularizing the SDK build however, it allows us to use
the jar stub source of one module as input to the SDK stubs metalava
run. This is sufficient to reproduce the sdk constant txt files, so that
we can avoid adding additional module SDK artifacts.

Bug: 187397780
Test: build module sdks, inspect for SdkConstant annotations
Test: diff framework-doc-stubs and framework-doc-stubs-module-stubs
      artifacts
Change-Id: I421b2e06cf406df4bab714b434b4dad947bee1f7