Don't reference non-public fields in annotations
Metalava would previously only consider whether a field was explicitly
hidden, and not whether it would have been visible if it wasn't hidden.

Note that this check is still incomplete. The full logic for whether a
symbol is visible or not is in ApiPredicate.test(), but that method
is not usable in this context as it results in an infinite loop. This
change however does fix an immediate problem of a hidden API being
included in the api txt.

Bug: 188527836
Test: m checkapi
Merged-In: I426f0b04e576e7411544d448937afaaf13f8a11a
Change-Id: I426f0b04e576e7411544d448937afaaf13f8a11a