Update ApiLookup construction to use the correct XML file
When the test was written, lint's API lookup method, when passed null as
the Android target, would fall back to calling
LintClient.findResource(...), which metalava was overriding to point it
to its intended XML file.

However, at some later point, lint's lookup code was made smarter, and
it will now find a suitable database to use, meaning that it doesn't
fall back to the findResource method, and as a result, metalava wasn't
actually consulting the newly constructed file, it was just consulting
the released SDK file. The new API Lookup construction method uses a
documented property to specifically point lint to the API database file
to use instead.

(Also replaces a deprecated usage of a lint API, as well as a deprecated
usage of a Guava API for creating a temporary directory.)

Test: Manually verified, as well as the DocAnalyzerTests (which were
  failing in my environment before this fix)
Change-Id: Ib4fdd91b546c10522be5097c321927c1a978cd6d