Add support for extracting a System API database for lint
190665366: Support @SystemApis in Android Lint in google3

Metalava constructs the api-versions.xml file used for lint's
NewApi check. This was written for the public SDK, so it assumes
that the first API level is 1. However, the System API starts
much later. This CL updates metava to make the starting API level
configurable; that way the build can be updated to also generate
the system api XML file needed by lint and we can start running
API checks on system apps as well.

Test: Unit test included
Bug: 190665366
Change-Id: Iae13aecaf47633b218d10701fa802edc32c0dac4