Improve lookup of parameter documentation
parameter.documentation() is always null because individual parameters
cannot be documented. This led to metalava often inserting null-related
documentation even though nullability was already discussed in the
original @param documentation.

Add some new utility methods for extracting the documentation that
relates to a single parameter, and use that to decide whether to insert
additional null documentation or not.

Note that this leads to many documentation changes in the SDK, usually
diffs of this sort:

   * @param callback the callback to add, must be non-null
-
-  * This value cannot be {@code null}.

Bug: 186197911
Test: unit tests in this CL
Test: inspect framework-doc-stubs output before & after
Merged-In: Ied2ed3158e143518ec3c3b61f88e1ddf9874f9f3
Change-Id: Ied2ed3158e143518ec3c3b61f88e1ddf9874f9f3