Add option to enhance documentation
Enhancing documentation is implied by --doc-stubs, but enabling it for
jar stubs is useful for making jar stubs more similar to the doc stubs.

Work is being done to modularize the "main" doc stubs build, such that
it no longer needs to see all the source code. The plan is to instead
provide it with the jar stub source of the modules. Enhancing the docs
of these jar stubs eliminates the need to keep many of the source
retention annotations that are used to enhance docs. It is likely we'll
be able to remove the "doc enhancement" of the doc stubs run, but for
now they are enhanced twice (harmless, as the annotations are not kept).

Bug: 186197911
Test: Other changes in this topic moves over the module stubs to use
this feature, and I've manually inspected the output of those. I've
also experimented with using these as input to framework-doc-stubs and
found that with the other recent changes to "null" detection in docs,
there is no effect on the output doc stubs.

Merged-In: I12280a1f62c2ced6f54fa28f71f72861ac86560a
Change-Id: I12280a1f62c2ced6f54fa28f71f72861ac86560a