Report issues at name declaration line
Previously, Reporter.elementToLocation found the first line of the PSI
element, which may have included a doc comment, making for confusing
line numbers. It had an includeDocs parameter which could be set to
false to disable this behavior, but it was unused. Disabling includeDocs
used the first line of the modifier list (if present) to as the
location. However, this also includes annotations which are frequently
on a line above the declaration.

This change attempts to improve the accuracy of elementToLocation by
finding the identifier element of the declaration, and computing its
line number. If one isn't present, it falls back to using the modifier
list and then to the parent element.

Test: Manually updated and verified line numbers in existing tests
Fix: 143153566
Change-Id: I1b7806604e81b14de6191021a2819c5b52f41d42