Modify @Discouraged to use class retention
This change modifies the AndroidX stub annotation to use class retention
instead of source retention, because it should not be stripped by the
compiler. This will allow the annotation to correctly pass through to
Android Studio.

This change also adds androidx.annotation.Discouraged to the hardcoded
list of annotations that is retained in all stubs.

Bug: 178477913
Test: - lunch sdk-eng; make sdk; javap -v -p -classpath
        out/host/linux-x86/sdk/sdk/android-sdk_eng.kevinjeon_linux-x86/
        platforms/android-31/android.jar android.graphics.Bitmap |
        grep Discouraged
      - Checked that the annotation appears in Bitmap's bytecode.
      - Checked that DiscouragedDetector correctly detects uses of
        annotated methods.
Change-Id: Ie2238b38e7b118c005e20ae48f53bbcfd8abfbcd