Improve lookup of @return documentation
Previously, when deciding whether to insert @return documentation for
nullability, metalava inspected the whole method documentation,
including parameter documentation. The result of this was that if a
parameter doc mentioned nullability, we'd sometimes skip inserting
@return nullability information.

Improve this so that we only look at the subset of the documentation
that would reasonably refer to the return value of the method.

Bug: 186197911
Test: tests in this cl
Test: diff framework-doc-stubs before & after
Change-Id: I15af628dc05d532aa67a12a374a4b1e0834b9167