Implicit nullability for annotation properties
Improve the implicit nullness check in PsiTypeItem to include
array-valued properties for Kotlin annotations. As in Java, the
components of such arrays must be non-null.

Additionally, add a check for the static modifier to avoid eargerly
applying the implict nullness to static methods/properties on
annotations or an instance method named `values` on enum classes.

Test: ApiFileTest
Fixes: 202968090
Change-Id: I221f578d9413fd80d6a3d54411b36c2e818cda94