Fix SdkConstant retention "detection"
The RewriteAnnotations class has different entry points, and its
determination of whether a particular annotation has source retention
or not only does a decent job in some of them.

When metalava is used in the --copy-annotations mode, it needs every
annotation that has source retention to be explicitly listed, so add
SdkConstant to that list.

Whilst we're here, make metalava fail more audibly (crash) if it
encounters an annotations with unknown retention, to ensure we catch
similar bugs sooner.

Bug: 225745567
Test: m private-stub-annotations ; inspect outputs (no SdkConstant.java)
Test: m android_stubs_current ; inspect jar (no SdkConstant.class)
Change-Id: I1ffaf954154562d8dd4f69a674a231b92e68f589