Filter out (some) extension SDK APIs when generating API
Teach metalava two new command line arguments:

  --sdk-jar-root: path to directory with prebuilt extension SDK jars
  --sdk-filter: path to filter file

The jar directory is expected to have a certain sub-folder setup, and is
usually $ANDROID_ROOT/prebuilts/sdk/extensions. Metalava will deduce
what APIs came from extension SDKs using these jars.

The filter file is a text file that determines which of the APIs from
the extension SDKs should be included in the complete API. [A future CL
will use the filter to set the extensions those APIs are part of (e.g.
method X is part of the S and T extensions).]

See Options.kt for more information on --sdk-jar-root and --sdk-filter.

Test: ./gradlew test --tests OptionsTest --tests 'ApiGeneratorTest*Extract*System*API'
Test: manual: diff api-versions.xml with and without --sdk-jar-root and --sdk-filter and verify APIs from extension SDKs not listed in the filter are removed
Bug: 228828986
Change-Id: I4d8b4457bdb5bf8e0404b3e2574c3819817e2ba5