Emit `from` attributes when generating API
Teach metalava to add the `from` attribute when generating an API. This
new attribute is used instead of the `since` attribute when an API
exists in one or more extension SDKs.

Note: for backwards compatibility, `since` and `from` can co-exist.
Existing tooling should be updated to ignore `since` in this case.

The format of the `from` attribute is

  from="ext:version[,ext:version[,...]]"

where <ext> is the numerical ID of the extension SDK, and <version> is
the numerical version of the extension SDK in which the API first
appeared.

See go/mainline-sdk-api-versions-xml for more information.

Also, bump `<api>` version to 3 to indicate the introduction of the
`from` attribute.

Bug: 228828986
Test: ./gradlew test --tests 'ApiGeneratorTest*Extract*System*API'
Change-Id: Id568ab4e7845f258206568d5bd82cf0c4d8c61de