sdk-extensions-info: accept extension SDKs, reject the platform SDK
The Android platform version in which an API first appeared is
calculated by comparing historical snapshots of platform SDKs. It is an
error to express the same information as part of the
sdk-extensions-info.

Throw an error if the info file lists the platform SDK in the list of
extension SDKs in which an API exists.

Bug: 245265483
Test: ./gradlew test --tests ApiToExtensionsMapTest
Change-Id: Iff2d408d42720de9851804efec552a22c983274e