Encode @sdkExtInfo javadoc comment in docstubs
Introduce a new javadoc comment @sdkExtInfo to hold information about
which SDK extensions and their respective versions an API first appeared
in.

Inject this javadoc comment in the docstubs that doclava consumes to
produce the final HTML documentation.

Implementation note: in order to generate the API documentation,
DocAnalyzer needs to include what SDK extensions (and the respective
versions) an API is available in. This information is available in the
api-versions.xml file that is passed to DocAnalyzer.applyApiLevels, but
the external framework used to parse the XML (lint-checks) does not
support SDK extensions info. Teach DocAnalyzer how to parse the SDK
extensions info from api-versions.xml. Future work may add that support
to the lint-checks library, and revert parts of this CL.

Bug: 228829064
Test: ./gradlew test --tests '*DocAnalyzerTest*Generate*SDK*'
Change-Id: I60694f109f450d1716a5def8cfb98cde8223ced5