Avoid using U/LC nullability annotation
It is a special representation in old frontend with lots of hacky utils
to compute type for nullability, qualified name, etc., which sometimes
led to @null nullability.

In FIR LC (or SLC) in the future, annotation abstraction (PsiAnnotation)
will be simplified/unified, and there is no more special class for
nullability.

For now, the best common way to determine if the given PsiAnnotation is
related to nullability is to check (the last part of) qualified name.

Bug: 229742431
Test: ./gradlew test
Change-Id: I7b7af6fa9ddd929bb335de1050a92a130eb1cd19