Respect the order of SDKs in sdk-extensions-info.txt
Each API symbol is related to a set of SDKs and the corresponding
versions the API first appeared in.

In order for Lint to provide the best suggestions for what runtime
checks to perform before calling an API in an extension SDK, the order
in which the SDKs are enumerated in api-verisons.xml is significant:
Lint will suggest the SDKs in that order.

Convert ApiToExtensionsMap from using a Set to using a List to preserve
the order the SDKs are listed in the sdk-extensions-info.txt file.

The Android SDK is not part of sdk-extensions-info.txt. As a special
case, always add the Android SDK after other SDKs.

Bug: 246772515
Test: ./gradlew test
Change-Id: I6e372221bd6dd0fa745c5ecca3dbefb8edcbb825