Refactor extension jar reading
Instead of creating several Api objects and merging them, use a single
Api object and feed all information into it from the JarReader.

To achieve this, refactor the jar reader to populate different fields
in the model when it's reading an extension jar vs an android jar.
Add fields in ApiElement to allow this.

Also move the "cleanup" methods calls of the Api object to after the Api
is fully populated.

See full diffs here: http://go/paste/4675966835097600

Bug: 258620828
Test: ./gradlew
Test: build api-versions.xml, find new extension classes and methods
      are now poopulated correctly. http://go/paste/4675966835097600
Change-Id: I926e38699a2c867e66d6432510e32f9dca74d634