Don't process jars that aren't in the allow-list
Most importantly, we currently put module APIs into the module-lib
and system-server api-versions.xml files if we do process these. It
would be better to tell metalava which jars to include depending
on which database we're generating, but this is a simpler/faster fix.

In theory there is a small performance win here as we don't scan some
jars unnecessarily.

Bug: 259115852
Bug: 258620828
Test: m api_versions_module_lib; verify no module APIs (except those
      in the sdk-ext-info-file.)
Change-Id: I3b53573ec85280d3d4331e60e94b0e9486ef44ab