Consolidate build type logic
If metalava is passed --current-codename <string>, it will operate in
"developer preview" mode: the internal API level will be the value of
--current-version <int> plus one, and the documentation for symbols
introduced in the upcoming codename dessert will include "Added in
<codename>" instead of "Added in <int>".

However, the special codename "REL" means the build is not a developer
preview after all. To summarize:

  - release build: codename is null or "REL"
  - developer preview build: all other cases

Teach Options to keep `currentCodeName` set to null if "REL" is passed
in. Add a helper function to Options so the rest of the codebase doesn't
have to know that `Options.currentCodeName == null` implies a developer
preview build.

Bug: 259367545
Test: ./gradlew test
Change-Id: Ibd252b96b8c306392fe24b540e241f3cd4505242