Workaround for return type nullability on @Deprecated
Such @Deprecated methods won't be created by LC, and UAST adds a fake
PSI instead. Return type nullability is added by LC, but UAST forgot to
do so while adding a fake one (KTIJ-23807).
Until the upstream fix is available, one workaround is to add a
nullability annotation if the given UMethod is a delegate to such fake
PSI and its return type nullability is available.

Bug: 219792969
Test: unit test modified
Change-Id: If8941a3ea8b306a7e30e947ebaa986979d87ffe1