Add flag to remove missing classes from inheritance entries
It's possible for a class in the non-updatable framework
to reference a class from an updatable apex module.
This currently happens with android.net.QosSessionAttributes,
which is an interface from the Connectivity module that
is implemented by a couple classes in frameworks/base/telephony.

Lint will complain and fail if it sees the <implements/> tag,
but not the actual class that's being implemented. So just
remove the <implements/> tag from the lint database.

Bug: 193460475
Test: m out/soong/.intermediates/frameworks/base/api_versions_module_lib/android_common/metalava/api-versions.xml, search for QosSessionAttributes before/after. Also ./gradlew
Change-Id: I4d80f2f8cb8a327f2219f2c2b829d1145f0ab2fe