Don't keep source retention annotations in tests
The option to include source retention annotations in the stubs is
currently only used in the tests. This means we're testing something
other than what we're actually running, so change the tests not to
use this option.

This enables removal of the the option to include source retention
annotations in a followup CL.

Bug: 230466482
Test: ./gradlew
Change-Id: I7ba9d0725a67668d78694988aa125966f2b156f0
(cherry picked from commit 92f48e56934975c52dc0495a71e0cc791a14c6be)