Add hasEqualReturnType() to TextClassItem class
The method compares two method items' return types and determines if the
two return types are considered equal in compiler level. This method is
used in the child CLs to determine the equality of the two method items.

Note that the method is not complete; Additional conditions for the
method will be introduced in the child CLs.

Test: ./gradlew
Bug: 269537741
Change-Id: I4a8b5d75c6b65ed07def0921b5b1743a29ee9c36