Add override equivalent signatures to text codebase
Currently, Metalava removes overridden methods in API text files if it
has equal signatures as the abstract method. This change adds method to
an interface when it inherits multiple interfaces with
override-equivalent signatures, as not resolving this leads to compile
time error.

Note that resolveAbstractMethods() is not complete with this change, as
this change only deals with interfaces. Completion of the method will be
accomplished in the child CL.

Test: ./gradlew
Bug: 269537741
Change-Id: I3ecca0daf89e2d890e878a2ca9969bf24d1499a2