Fix test/build failures caused by update to 31.1.0-alpha01
The failures are due to changes in the linter code to properly support
SDK extensions (see ag/18185655) which causes the `ApiLookup` code to
stop using the `since` attribute and use the `sdks` attribute instead.
Unsurprisingly, the failures are in the SDK extension specific tests.

* Build failures are due to deprecation of `ApiLookup.get*Version(...)`
  methods. They were deprecated in favor of `get*Versions(...)` methods
  that return an `ApiConstraint` instead of an int version. That is
  needed because the introduction of SDK extensions means that there is
  no longer a single version when an API was added.
* Test failures caused by differences in the results caused by use of
  the `sdks` attribute instead of the `since` attribute.

The tests and the code need to be updated now that the linter properly
supports SDK extensions. There appears to be some overlap between code
that uses `SdkAndVersion` and the new code added in `ApiLookup` and
`ApiConstraint`. The tests also need to set the `sdks` attribute so
that it is more consistent with the `since` attribute and switch to
using non-deprecated methods. However, that is outside the scope of
this bug and doing that work before completing the de-fork will likely
introduce more conflicts and make de-forking harder.

There appears to be a bug in the implementation of
`ApiConstraint.min()` which is an alias for
`ApiConstraint.fromInclusive()`. See b/282932318.

The fixes that are made in this change are:
* Adding an ApiConstraint.minApiLevel() method that implements the
  correct behavior for `ApiConstraint.fromInclusive()/min()`.
* Use the new `minApiLevel()` method instead of relying on deprecated
  methods that use the broken `min()` method.
* Added some information that was missing from the `sdks` attribute
  in one test.

Bug: 281035030
Test: ./gradlew
      m update-api
Change-Id: I0dc7e11d4e23211ea38df6e2ac12c2b90705cc22