Update to AGP 8.2.0-alpha03
...which includes UAST 231 artifacts
that has fixes for type/nullity of inline/deprecated-hidden function.
Most of workarounds for KTIJ-23807 and KTIJ-23837 are gone which are
added at b/219792969, b/240475358, b/244408981, and b/238733178.
One thing we still need to keep is to add nullity annotation to the
element type of last vararg parameter's array type, which is still
missed in UAST's fake PSI.

In addition, it improves:
* upper bound wildcards for return type of inline/deprecated-hidden
* nullity of implicit return type of inline/deprecated-hidden

Bug: 262033498
Bug: 271032477
Test: ./gradlew test
Change-Id: Ia0ac8605b59dc61dbe15a2266b11fc29c61d69ac