Really use absolute paths for kotlin files
Change https://r.android.com/2594328 tried to use absolute paths for
kotlin files and it did but in the process of refactoring between PS#2
and PS#3 part of the change was regressed so it no longer used absolute
files for the source roots. It is not clear why that did not cause
downstream build failures but other changes being merged in from
`metalava-main` do cause failures due to that regression so this change
fixes that regression.

Bug: 281035030
Test: ./gradlew
      m checkapi
Change-Id: If6935cf8b928a8eff7454febac925a041eb52a01