Add REMOVED_FINAL_STRICT issue
Previously, the compatibility issues caused by removal of `final` from
a method was tracked using `REMOVED_FINAL`. The compatibility issues
arise when you build against an API without `final` that allows the
method to be overridden but is run against an older implementation that
still has the `final` modifier. That is not an issue for AndroidX where
the method is statically included in the App.

However, re-using the `REMOVED_FINAL` for this prevents AndroidX from
disabling the check without losing valuable compatibility checks that
do affect it.

This creates a new `REMOVED_FINAL_STRICT` issue to track removal of
`final` from a method.

Bug: 233880757
Test: ./gradlew
      m checkapi
Change-Id: I62a4be01df8afa8d466ad204582b3d7731f31f18