Remove unused --only-check-api/--only-update-api
The purpose of these was to avoid doing a lot of unnecessary work when
updating and checking APIs. It is not clear whether they were ever used
but at the moment they appear to be unused. Similar, albeit cleaner,
functionality can be added as a separate `check-api` / `update-api`
sub-command if necessary.

Bug: 286047790
Test: ./gradlew
      m checkapi - in AOSP master
      grepped for usage in Google3/AndroidX.
Change-Id: I8456d79025c0e692dd526ef7efd0b76c4bf93016