Add a new version sub-command
This adds a simple `version` sub-command that outputs the metalava
version, similar to the `--version` option.

Adding the first sub-command changes the default behavior of a Clikt
command and requires some additional integration work to ensure that
they work consistently with the legacy (no sub-command) metalava
command. By adding a simple `version` command this minimizes the
size of this change.

The main changes in addition to adding the `version` sub-command are:
* Updating `MetalavaCommand` to handle sub-commands.
* Adding metalava specific customizations (i.e. formatting and
  localization) of standard Clikt help.
* Adding a `MetalavaSubCommand` base for all metalava sub-commands.

Bug: 286023667
Test: ./gradlew --rerun-tasks
      ./gradlew installDist
      export PATH=/usr/lib/jvm/java-17-openjdk-amd64/bin:$PATH
      ../../out/metalava/install/metalava/bin/metalava version
      ../../out/metalava/install/metalava/bin/metalava version -h
Change-Id: I7cdb08f680f5bd9650047ddbde3e29905424965c