Remove dependency from Codebase onto asm classes
While depending on asm classes would not prevent separating the model
package into its own library the methods that use those classes are
only used in `ConvertJarsToSignatureFiles` and so are not generally
useful. As they only depend on other methods exported by the model it
was simple to just convert them into extension functions.

Bug: 287241408
Test: ./gradlew
Change-Id: Ied909dde46f91dba31cf68e1bc17cadc0354160a