Extract resolving functions into ReferenceResolver class
Resolving the references in a `TextCodebase` is actually part of the
parsing process rather than the core functionality of `TextCodebase`.
This change extracts the functionality into a nested class and moves
some of the parameters into state in the `ReferenceResolver`. That is
in preparation for moving the logic into `ApiFile.kt`.

Bug: 288550547
Test: ./gradlew
Change-Id: Ie26192ac28dd9e785fe801b2e6a9d0f6e979916f